+++ src/modulos/rol-to-user/controller/UserRolController.ts
@@ -9,24 +9,25 @@ export class UserRolController {
 
 
     async userRol(request: Request, response: Response, next: NextFunction): Promise<GenericResponse> {
-        // console.log('method userRol');
+        console.log('method userRol');
         let resp: GenericResponse = new GenericResponse();
         let registroToRemove: UserRol[] = [];
+        let isDelete: boolean = false;
         const {
-            listRoles
+            listRolesId
         } = request.body;
-        // console.log(JSON.stringify(listRoles));
+        // console.log('listRolesId: ' + JSON.stringify(listRolesId));
         // console.log('Paso Uno');
         try {
             const idUser = parseInt(request.params.iduser);
+            // console.log('idUser: ' + idUser);
             registroToRemove = await this.repository.find({ where: { userId: idUser } });
-            // console.log(JSON.stringify(registroToRemove));
+            // console.log('registroToRemove: ' + JSON.stringify(registroToRemove));
             if (registroToRemove !== null
                 && typeof registroToRemove !== 'undefined'
-                && registroToRemove.length === 0) {
-                resp.code = '-4';
-                resp.message = StatusCode.ERROR + ': Relación de Roles no existe';
-                return resp;
+                && registroToRemove.length > 0) {
+                isDelete = true;
+                console.log('Tiene Roles asignados para eliminar');
             }
         } catch (error) {
             resp.code = '-3';
@@ -35,26 +36,30 @@ export class UserRolController {
             return resp;
         }
         // console.log('Paso Dos');
-        try {
-            const removeVal: any = await this.repository.remove(registroToRemove);
-        } catch (error) {
-            // console.log(JSON.stringify(error));
-            resp.code = '-2';
-            resp.message = StatusCode.ERROR;
-            resp.data = null;
+        if (isDelete) {
+            try {
+                const removeVal: any = await this.repository.remove(registroToRemove);
+                console.log('Roles existentes Eliminados');
+            } catch (error) {
+                // console.log(JSON.stringify(error));
+                resp.code = '-2';
+                resp.message = StatusCode.ERROR;
+                resp.data = null;
+            }
         }
         // console.log('Paso Tres');
         try {
             const idUser = parseInt(request.params.iduser);
             let saveUserRol: UserRol[] = [];
-            for (let index = 0; index < listRoles.length; index++) {
+            for (let index = 0; index < listRolesId.length; index++) {
                 let item: UserRol = new UserRol();
                 item.userId = idUser;
-                item.rolId = listRoles[index];;
+                item.rolId = listRolesId[index];;
                 saveUserRol.push(item);
             }
             // console.log(JSON.stringify(saveUserRol));
             const removeVal: any = await this.repository.save(saveUserRol);
+            console.log('Nuevo set de roles asignados correctamente');
             resp.data = null;
         } catch (error) {
             // console.log(JSON.stringify(error));
